+++ gst-libs/gst/vaapi/gstvaapiwindow_x11.c
@@ -626,7 +626,7 @@ gst_vaapi_window_x11_new(GstVaapiDisplay *display, guint width, guint height)
 GstVaapiWindow *
 gst_vaapi_window_x11_new_with_xid(GstVaapiDisplay *display, Window xid)
 {
-    GST_DEBUG("new window from xid 0x%08x", xid);
+    GST_DEBUG("new window from xid 0x%08x", (guint)xid);
 
     g_return_val_if_fail(GST_VAAPI_IS_DISPLAY_X11(display), NULL);
     g_return_val_if_fail(xid != None, NULL);
